Deployement v1

+++ back-end/src/Customer/Customer.controller.ts
@@ -8,7 +8,9 @@ import {
   Get,
   Delete,
   Put,
+  Request,
 } from '@nestjs/common';
+import { Request as ExpressRequest } from 'express';
 import { CustomerDTO, UpdateCustomer } from './Customer.DTO';
 import { CustomerService } from './Customer.service';
 import { AdminAuthGuard } from '../Authentication/Authentication.AdminAuthgurd';
@@ -19,16 +21,36 @@ export class CustomerController {
 
   @UseGuards(AdminAuthGuard)
   @Post('Customer_register')
-  async registerCustomer(@Body() customerDto: CustomerDTO, @Req() req) {
-    const AdminId = req.user.userId;
-    return this.customerService.registedCustomer(customerDto, AdminId);
+  async registerCustomer(
+    @Body() customerDto: CustomerDTO,
+    @Req() req: ExpressRequest & { user?: { userId: string } },
+  ) {
+    const AdminId = req.user?.userId;
+    if (!AdminId) {
+      throw new Error('AdminId is missing');
+    }
+    return this.customerService.registedCustomer(
+      customerDto,
+      parseInt(AdminId, 10),
+    );
   }
 
   @UseGuards(AdminAuthGuard)
   @Get('getAllCustomers')
-  async getAllCustomers() {
-    return this.customerService.getAllCustomers();
+  async getAllCustomers(
+    @Query('page') page: string,
+    @Query('limit') limit: string,
+  ) {
+    return this.customerService.getAllCustomers(
+      Number(page) || 1,
+      Number(limit) || 10,
+    );
   }
+
+  // @Get('getAllCustomers')
+  // async getAllCustomers() {
+  //   return this.customerService.getAllCustomers();
+  // }
 
   @Post('verifyCustomer')
   async verifyCustomer(@Query('customerId') customerId: string) {
@@ -37,18 +59,30 @@ export class CustomerController {
 
   @UseGuards(AdminAuthGuard)
   @Delete('deleteCustomer')
-  async deleteCustomer(@Query('customerId') customerId: string, @Req() req) {
-    const AdminId = req.user.userId;
+  async deleteCustomer(
+    @Query('customerId') customerId: string,
+    @Req() req: ExpressRequest & { user?: { userId: string } },
+  ) {
+    const AdminId = req.user?.userId;
+    if (!AdminId) {
+      throw new Error('AdminId is missing');
+    }
     return this.customerService.removeCustomer(
       parseInt(customerId, 10),
-      AdminId,
+      parseInt(AdminId, 10),
     );
   }
 
   @UseGuards(AdminAuthGuard)
   @Put('updateCustomer')
-  async updateCustomer(@Body() UpdateCustomer: UpdateCustomer, @Req() req) {
-    const AdminId = req.user.userId;
+  async updateCustomer(
+    @Body() UpdateCustomer: UpdateCustomer,
+    @Req() req: ExpressRequest & { user?: { userId: string } },
+  ) {
+    const AdminId = req.user?.userId;
+    if (!AdminId) {
+      throw new Error('AdminId is missing');
+    }
     return this.customerService.updateCustomer(
       UpdateCustomer,
       parseInt(AdminId, 10),
